Test cpu jitter entropy (#5)
* Add CPU Jitter as the entropy source for FIPS build (#213)

* Add unit tests for CPU Jitter and re-enable urandom_test (#242)

Add unit tests for CPU Jitter and re-enable urandom_test (#242)

Co-authored-by: Dusan Kostic <dkostic@amazon.com>

* Correctly check Jitter read_entropy return value (#248)

Documentation for Jitter read_entropy functions states that the functions return either a negative integer which signifies an error, or the number of generated bytes of entropy. The implementation of the functions however checks if it the requested and the generated number of bytes is the same and returns an error if it's not. Nevertheless, to be on the safe side we now check that Jitter function returns the expected value.

Co-authored-by: Dusan Kostic <dkostic@amazon.com>

* Remove unused macro check in urandom_test (#255)

Co-authored-by: Dusan Kostic <dkostic@amazon.com>

* Add third_party/jitterentropy license to the top level LICENSE file (#266)

CryptoAlg-863 V438103860

Co-authored-by: dkostic <25055813+dkostic@users.noreply.github.com>
Co-authored-by: Dusan Kostic <dkostic@amazon.com>